Written a benchmark for trainers.
The steps count is very low, 10 per trainer, while most trainers are
meant to be run with at least a hundred steps. This puts too much
emphasis on the trainer code, but it takes quite a while already, so
that's a compromise; if unsure whether something really matters
increase the count temporarily, or run an actual training session.

Both bots and trainers could also have proper tests (for example
checking the results of a playthrough or a training session with a
fixed prngs seed).