Made the multi-valuedness explicit and shape checking stricter.
The base bot supports using different parameters at different parts of
the level almost out-of-the-box (an _m bot variation has to be written
-- but that requires just adding one dimension to each array and using
a base-managed index for it).

Further, we want to handle all 3 transforms of seed shapes:
* single-choice parameters replicated to a n-choices set (e.g. using
  a linear seed for the linear_m bot);
* multi-choice parameters clipped to a single-set (e.g. using a
  linear_m seed for the linear bot);
* m-choice parameters "adapted" for n-choices (e.g. using a linear_m
  seed with 2 choices for the linear_m bot with 3 phases).

Now _m bots are required to set self.param_multi, but in exchange a
degenerate case of a multi-set with a single choice works correctly
and the parameter shape check can catch most mismatches.