Coded a processor that can detect level subepisodes.
This new processor should be able to discover level splits if they are
marked by a state component being reset to its initial value. This was
inspired by the description of the level structure from Victor's paper:
http://research.nextmind.org/publications/BackBox_Challenge_Ru.pdf.

The train level (but not the test level) also seems to have 3 more or
less equally spaced resets of the 3rd component (thus 4 subepisodes;
apparent for both random walks and good bots). On the other hand, the
9 resets of the last component are pretty clear for random walks on
both levels, but not so clear with good bots.